Swap `unsigned long` with `size_t`
- Throughout the program, there are instances of "unsigned long" and "size_t"
- Under Mac and Linux, "unsigned long" is synonymous with "size_t"

- Replace all instances of "unsigned long" with "size_t"
- The function `sw_strtoul()` now returns "size_t"
	* Update documentation to include a note about "size_t"
	* This implementation is currently assuming "size_t" will be used on Linux or Mac

Important notes
- Windows implementations of "size_t" will most likely make it "unsigned long long" which is 64 bits (32 bits on Mac/Linux)
- If this is to be used on Windows or a lower-bit system, `sw_strtoul()` may need to be modified

+++ include/SW_Control.h
@@ -41,7 +41,7 @@ void SW_RUN_deepCopy(
 
 void SW_CTL_setup_domain(
     int rank,
-    unsigned long userSUID,
+    size_t userSUID,
     Bool renameDomainTemp,
     SW_DOMAIN *SW_Domain,
     LOG_INFO *LogInfo
@@ -87,11 +87,11 @@ void SW_CTL_run_current_year(SW_RUN *sw, SW_OUT_DOM *OutDom, LOG_INFO *LogInfo);
 void SW_CTL_run_spinup(SW_RUN *sw, SW_OUT_DOM *OutDom, LOG_INFO *LogInfo);
 
 void SW_CTL_run_sw(
-    unsigned long runNum,
+    size_t runNum,
     SW_RUN_INPUTS *runInputs,
     SW_RUN *sw_template,
     SW_DOMAIN *SW_Domain,
-    unsigned long ncSuid[],
+    size_t ncSuid[],
     Bool estVeg,
     Bool copyWeather,
     const size_t count[],
